[APP] Add and refactor user association with student number
