+++ mon-pix/app/components/routes/campaigns/restricted/join-sco.js
@@ -1,0 +1,143 @@
+import { action } from '@ember/object';
+import { inject as service } from '@ember/service';
+import { tracked } from '@glimmer/tracking';
+import Component from '@glimmer/component';
+import { standardizeNumberInTwoDigitFormat } from 'mon-pix/utils/standardize-number';
+
+const ERROR_INPUT_MESSAGE_MAP = {
+  firstName: 'Votre prénom n’est pas renseigné.',
+  lastName: 'Votre nom n’est pas renseigné.',
+  dayOfBirth: 'Votre jour de naissance n’est pas valide.',
+  monthOfBirth: 'Votre mois de naissance n’est pas valide.',
+  yearOfBirth: 'Votre année de naissance n’est pas valide.',
+};
+
+const isDayValid = (value) => value > 0 && value <= 31;
+const isMonthValid = (value) => value > 0 && value <= 12;
+const isYearValid = (value) => value > 999 && value <= 9999;
+const isStringValid = (value) => !!value.trim();
+
+class Validation {
+  @tracked firstName = null
+  @tracked lastName = null
+  @tracked dayOfBirth = null
+  @tracked monthOfBirth = null
+  @tracked yearOfBirth = null
+}
+
+export default class JoinSco extends Component {
+  queryParams = ['participantExternalId'];
+  participantExternalId = null;
+
+  @service session;
+  @service currentUser;
+  @service store;
+
+  validation = new Validation();
+
+  @tracked firstName = '';
+  @tracked lastName = '';
+  @tracked dayOfBirth = '';
+  @tracked monthOfBirth = '';
+  @tracked yearOfBirth = '';
+  @tracked studentNumber = '';
+
+  constructor() {
+    super(...arguments);
+
+    if (this.session.data.authenticated.source === 'external') {
+      this.firstName = this.currentUser.user.firstName;
+      this.lastName = this.currentUser.user.lastName;
+    }
+  }
+
+  get birthdate() {
+    const dayOfBirth = standardizeNumberInTwoDigitFormat(this.dayOfBirth.trim());
+    const monthOfBirth = standardizeNumberInTwoDigitFormat(this.monthOfBirth.trim());
+    const yearOfBirth = this.yearOfBirth.trim();
+    return [yearOfBirth, monthOfBirth, dayOfBirth].join('-');
+  }
+
+  get isFormNotValid() {
+    return !isStringValid(this.firstName) || !isStringValid(this.lastName)
+      || !isDayValid(this.dayOfBirth) || !isMonthValid(this.monthOfBirth) || !isYearValid(this.yearOfBirth);
+  }
+
+  get isDisabled() {
+    return this.session.data.authenticated.source === 'external';
+  }
+
+  @action
+  attemptNext() {
+    this._validateInputName('firstName', this.firstName);
+    this._validateInputName('lastName', this.lastName);
+    this._validateInputDay('dayOfBirth', this.dayOfBirth);
+    this._validateInputMonth('monthOfBirth', this.monthOfBirth);
+    this._validateInputYear('yearOfBirth', this.yearOfBirth);
+
+    const schoolingRegistration = this.store.createRecord('schooling-registration-user-association', {
+      id: this.args.campaignCode + '_' + this.lastName,
+      firstName: this.firstName,
+      lastName: this.lastName,
+      birthdate: this.birthdate,
+      campaignCode: this.args.campaignCode,
+    });
+    if (!this.isFormNotValid) {
+      return this.args.onSubmit(schoolingRegistration);
+    }
+  }
+
+  @action
+  triggerInputStringValidation(key, value) {
+    this._validateInputName(key, value);
+  }
+
+  @action
+  triggerInputDayValidation(key, value) {
+    value = value.trim();
+    this._standardizeNumberInInput('dayOfBirth', value);
+    this._validateInputDay(key, value);
+  }
+
+  @action
+  triggerInputMonthValidation(key, value) {
+    value = value.trim();
+    this._standardizeNumberInInput('monthOfBirth', value);
+    this._validateInputMonth(key, value);
+  }
+
+  @action
+  triggerInputYearValidation(key, value) {
+    value = value.trim();
+    this.yearOfBirth = value;
+    this._validateInputYear(key, value);
+  }
+
+  _executeFieldValidation(key, value, isValid) {
+    const isInvalidInput = !isValid(value);
+    const message = isInvalidInput ? ERROR_INPUT_MESSAGE_MAP[key] : null;
+    this.validation[key] = message;
+  }
+
+  _validateInputName(key, value) {
+    this._executeFieldValidation(key, value, isStringValid);
+  }
+
+  _validateInputDay(key, value) {
+    this._executeFieldValidation(key, value, isDayValid);
+  }
+
+  _validateInputMonth(key, value) {
+    this._executeFieldValidation(key, value, isMonthValid);
+  }
+
+  _validateInputYear(key, value) {
+    this._executeFieldValidation(key, value, isYearValid);
+  }
+
+  _standardizeNumberInInput(attribute, value) {
+    if (value) {
+      this.attribute = standardizeNumberInTwoDigitFormat(value);
+    }
+  }
+}
