+++ mon-pix/app/controllers/campaigns/restricted/join.js
@@ -3,29 +3,7 @@ import { action } from '@ember/object';
 import { inject as service } from '@ember/service';
 import { tracked } from '@glimmer/tracking';
 import Controller from '@ember/controller';
-import { standardizeNumberInTwoDigitFormat } from 'mon-pix/utils/standardize-number';
 import ENV from 'mon-pix/config/environment';
-
-const ERROR_INPUT_MESSAGE_MAP = {
-  firstName: 'Votre prénom n’est pas renseigné.',
-  lastName: 'Votre nom n’est pas renseigné.',
-  dayOfBirth: 'Votre jour de naissance n’est pas valide.',
-  monthOfBirth: 'Votre mois de naissance n’est pas valide.',
-  yearOfBirth: 'Votre année de naissance n’est pas valide.',
-};
-
-const isDayValid = (value) => value > 0 && value <= 31;
-const isMonthValid = (value) => value > 0 && value <= 12;
-const isYearValid = (value) => value > 999 && value <= 9999;
-const isStringValid = (value) => !!value.trim();
-
-class Validation {
-  @tracked firstName = null
-  @tracked lastName = null
-  @tracked dayOfBirth = null
-  @tracked monthOfBirth = null
-  @tracked yearOfBirth = null
-}
 
 export default class JoinRestrictedCampaignController extends Controller {
   queryParams = ['participantExternalId'];
@@ -35,58 +13,17 @@ export default class JoinRestrictedCampaignController extends Controller {
   @service store;
   @service intl;
 
-  validation = new Validation();
-
   @tracked isLoading = false;
   @tracked errorMessage = null;
 
-  @tracked firstName = '';
-  @tracked lastName = '';
-  @tracked dayOfBirth = '';
-  @tracked monthOfBirth = '';
-  @tracked yearOfBirth = '';
-  @tracked studentNumber = '';
-
-  get birthdate() {
-    const dayOfBirth = standardizeNumberInTwoDigitFormat(this.dayOfBirth.trim());
-    const monthOfBirth = standardizeNumberInTwoDigitFormat(this.monthOfBirth.trim());
-    const yearOfBirth = this.yearOfBirth.trim();
-    return [yearOfBirth, monthOfBirth, dayOfBirth].join('-');
-  }
-
-  get isFormNotValid() {
-    return !isStringValid(this.firstName) || !isStringValid(this.lastName)
-    || !isDayValid(this.dayOfBirth) || !isMonthValid(this.monthOfBirth) || !isYearValid(this.yearOfBirth);
-  }
-
-  get isDisabled() {
-    return this.session.data.authenticated.source === 'external';
-  }
-
   @action
-  attemptNext() {
+  attemptNext(schoolingRegistration) {
+
     this.errorMessage = null;
     this.isLoading = true;
-    this._validateInputName('firstName', this.firstName);
-    this._validateInputName('lastName', this.lastName);
-    this._validateInputDay('dayOfBirth', this.dayOfBirth);
-    this._validateInputMonth('monthOfBirth', this.monthOfBirth);
-    this._validateInputYear('yearOfBirth', this.yearOfBirth);
-    if (this.isFormNotValid) {
-      return this.isLoading = false;
-    }
-
-    const campaignCode = this.model.code;
-    const schoolingRegistration = this.store.createRecord('schooling-registration-user-association', {
-      id: campaignCode + '_' + this.lastName,
-      firstName: this.firstName,
-      lastName: this.lastName,
-      birthdate: this.birthdate,
-      campaignCode,
-    });
-
     return schoolingRegistration.save().then(() => {
       this.isLoading = false;
+
       this.transitionToRoute('campaigns.start-or-resume', this.model.code, {
         queryParams: { associationDone: true, participantExternalId: this.participantExternalId }
       });
@@ -95,32 +32,6 @@ export default class JoinRestrictedCampaignController extends Controller {
       this._setErrorMessageForAttemptNextAction(errorResponse);
       this.isLoading = false;
     });
-  }
-
-  @action
-  triggerInputStringValidation(key, value) {
-    this._validateInputName(key, value);
-  }
-
-  @action
-  triggerInputDayValidation(key, value) {
-    value = value.trim();
-    this._standardizeNumberInInput('dayOfBirth', value);
-    this._validateInputDay(key, value);
-  }
-
-  @action
-  triggerInputMonthValidation(key, value) {
-    value = value.trim();
-    this._standardizeNumberInInput('monthOfBirth', value);
-    this._validateInputMonth(key, value);
-  }
-
-  @action
-  triggerInputYearValidation(key, value) {
-    value = value.trim();
-    this.yearOfBirth = value;
-    this._validateInputYear(key, value);
   }
 
   _setErrorMessageForAttemptNextAction(errorResponse) {
@@ -149,33 +60,5 @@ export default class JoinRestrictedCampaignController extends Controller {
     ];
 
     return  _.find(errors, ['shortCode', meta.shortCode]).message || defaultMessage;
-  }
-
-  _executeFieldValidation(key, value, isValid) {
-    const isInvalidInput = !isValid(value);
-    const message = isInvalidInput ? ERROR_INPUT_MESSAGE_MAP[key] : null;
-    this.validation[key] = message;
-  }
-
-  _validateInputName(key, value) {
-    this._executeFieldValidation(key, value, isStringValid);
-  }
-
-  _validateInputDay(key, value) {
-    this._executeFieldValidation(key, value, isDayValid);
-  }
-
-  _validateInputMonth(key, value) {
-    this._executeFieldValidation(key, value, isMonthValid);
-  }
-
-  _validateInputYear(key, value) {
-    this._executeFieldValidation(key, value, isYearValid);
-  }
-
-  _standardizeNumberInInput(attribute, value) {
-    if (value) {
-      this.attribute = standardizeNumberInTwoDigitFormat(value);
-    }
   }
 }
