+++ mon-pix/app/routes/campaigns/restricted/join.js
@@ -14,7 +14,7 @@ export default class JoinRoute extends Route.extend(SecuredRouteMixin) {
   async redirect(campaign) {
     let schoolingRegistration = await this.store.queryRecord('schooling-registration-user-association', { userId: this.currentUser.user.id, campaignCode: campaign.code });
 
-    if (isEmpty(schoolingRegistration)) {
+    if (isEmpty(schoolingRegistration) && campaign.organizationType === 'SCO') {
       schoolingRegistration = await this.store.createRecord('schooling-registration-user-association', { userId: this.currentUser.user.id, campaignCode: campaign.code }).save({ adapterOptions: { tryReconciliation: true } });
     }
