+++ mon-pix/tests/unit/controllers/campaigns/restricted/join-test.js
@@ -9,482 +9,45 @@ describe('Unit | Controller | campaigns/restricted/join', function() {
   let controller;
   let storeStub;
   let sessionStub;
-  let studentUserAssociation;
+  let schoolingRegistration;
 
   beforeEach(function() {
     controller = this.owner.lookup('controller:campaigns.restricted.join');
     controller.transitionToRoute = sinon.stub();
-    studentUserAssociation = { save: sinon.stub(), unloadRecord: sinon.stub() };
-    const createStudentUserAssociationStub = sinon.stub().returns(studentUserAssociation);
-    storeStub = { createRecord: createStudentUserAssociationStub };
+    schoolingRegistration = { save: sinon.stub(), unloadRecord: sinon.stub() };
+    const createschoolingRegistrationStub = sinon.stub().returns(schoolingRegistration);
+    storeStub = { createRecord: createschoolingRegistrationStub };
     sessionStub = { data: { authenticated: { source: 'pix' } } };
     controller.set('store', storeStub);
     controller.set('session', sessionStub);
     controller.set('model', 'AZERTY999');
   });
 
-  describe('#triggerInputDayValidation', function() {
-    context('when dayOfBirth is invalid', function() {
-
-      [
-        '',
-        ' ',
-        '32',
-        '0',
-        '444',
-        'ee',
-      ].forEach(function(wrongDayOfBirth) {
-        it(`should display an error when dayOfBirth is ${wrongDayOfBirth}`, async function() {
-          // when
-          await controller.actions.triggerInputDayValidation.call(controller, 'dayOfBirth', wrongDayOfBirth);
-
-          // then
-          expect(controller.validation.dayOfBirth).to.equal('Votre jour de naissance n’est pas valide.');
-        });
-      });
-    });
-
-    context('when dayOfBirth is valid', function() {
-
-      [
-        '1',
-        '01',
-        '31',
-      ].forEach(function(validDayOfBirth) {
-        it(`should not display an error when dayOfBirth is ${validDayOfBirth}`, async function() {
-          // when
-          await controller.actions.triggerInputDayValidation.call(controller, 'dayOfBirth', validDayOfBirth);
-
-          // then
-          expect(controller.get('validation.dayOfBirth')).to.equal(null);
-        });
-      });
-    });
-  });
-
-  describe('#triggerInputMonthValidation', function() {
-
-    context('when monthOfBirth is invalid', function() {
-
-      [
-        '',
-        ' ',
-        '13',
-        '0',
-        '444',
-        'ee',
-      ].forEach(function(wrongMonthOfBirth) {
-        it(`should display an error when monthOfBirth is ${wrongMonthOfBirth}`, async function() {
-          // when
-          await controller.actions.triggerInputMonthValidation.call(controller, 'monthOfBirth', wrongMonthOfBirth);
-
-          // then
-          expect(controller.get('validation.monthOfBirth')).to.equal('Votre mois de naissance n’est pas valide.');
-        });
-      });
-    });
-
-    context('when monthOfBirth is valid', function() {
-
-      [
-        '1',
-        '01',
-        '12',
-      ].forEach(function(validMonthOfBirth) {
-        it(`should not display an error when monthOfBirth is ${validMonthOfBirth}`, async function() {
-          // when
-          await controller.actions.triggerInputMonthValidation.call(controller, 'monthOfBirth', validMonthOfBirth);
-
-          // then
-          expect(controller.get('validation.monthOfBirth')).to.equal(null);
-        });
-      });
-    });
-  });
-
-  describe('#triggerInputYearValidation', function() {
-
-    context('when yearOfBirth is invalid', function() {
-
-      [
-        '',
-        ' ',
-        '1',
-        '11',
-        '100',
-        '0000',
-        '0001',
-        '0011',
-        '0111',
-        '10000',
-      ].forEach(function(wrongYearOfBirth) {
-        it(`should display an error when yearOfBirth is ${wrongYearOfBirth}`, async function() {
-          // when
-          await controller.actions.triggerInputYearValidation.call(controller, 'yearOfBirth', wrongYearOfBirth);
-
-          // then
-          expect(controller.get('validation.yearOfBirth')).to.equal('Votre année de naissance n’est pas valide.');
-        });
-      });
-    });
-
-    context('when yearOfBirth is valid', function() {
-
-      [
-        '1000',
-        '9999',
-      ].forEach(function(validYearOfBirth) {
-        it(`should not display an error when yearOfBirth is ${validYearOfBirth}`, async function() {
-          // when
-          await controller.actions.triggerInputYearValidation.call(controller, 'yearOfBirth', validYearOfBirth);
-
-          // then
-          expect(controller.get('validation.yearOfBirth')).to.equal(null);
-        });
-      });
-    });
-  });
-
-  describe('#triggerInputStringValidation', function() {
-
-    context('when string is invalid', function() {
-
-      [
-        '',
-        ' ',
-      ].forEach(function(wrongString) {
-        it(`should display an error when firstName is "${wrongString}"`, async function() {
-          // when
-          await controller.actions.triggerInputStringValidation.call(controller, 'firstName', wrongString);
-
-          // then
-          expect(controller.get('validation.firstName')).to.equal('Votre prénom n’est pas renseigné.');
-        });
-
-        it(`should display an error when lastName is "${wrongString}"`, async function() {
-          // when
-          await controller.actions.triggerInputStringValidation.call(controller, 'lastName', wrongString);
-
-          // then
-          expect(controller.get('validation.lastName')).to.equal('Votre nom n’est pas renseigné.');
-        });
-      });
-    });
-
-    context('when string is valid', function() {
-
-      [
-        'Robert',
-        'Smith',
-      ].forEach(function(validString) {
-        it(`should not display an error when firstName is ${validString}`, async function() {
-          // when
-          await controller.actions.triggerInputStringValidation.call(controller, 'firstName', validString);
-
-          // then
-          expect(controller.get('validation.firstName')).to.equal(null);
-        });
-
-        it(`should not display an error when lastName is ${validString}`, async function() {
-          // when
-          await controller.actions.triggerInputStringValidation.call(controller, 'lastName', validString);
-
-          // then
-          expect(controller.get('validation.lastName')).to.equal(null);
-        });
-      });
-    });
-  });
-
-  describe('#isFormNotValid', function() {
-
-    it('should be true if firstName is not valid', function() {
-      // given
-      controller.set('firstName', ' ');
-      controller.set('lastName', 'Smith');
-      controller.set('dayOfBirth', '15');
-      controller.set('monthOfBirth', '12');
-      controller.set('yearOfBirth', '1999');
-
-      // when
-      const result = controller.get('isFormNotValid');
-
-      // then
-      expect(result).to.equal(true);
-    });
-
-    it('should be true if lastName is not valid', function() {
-      // given
-      controller.set('firstName', 'Robert');
-      controller.set('lastName', '');
-      controller.set('dayOfBirth', '15');
-      controller.set('monthOfBirth', '12');
-      controller.set('yearOfBirth', '99999');
-
-      // when
-      const result = controller.get('isFormNotValid');
-
-      // then
-      expect(result).to.equal(true);
-    });
-
-    it('should be true if dayOfBirth is not valid', function() {
-      // given
-      controller.set('dayOfBirth', '99');
-      controller.set('monthOfBirth', '12');
-      controller.set('yearOfBirth', '1999');
-
-      // when
-      const result = controller.get('isFormNotValid');
-
-      // then
-      expect(result).to.equal(true);
-    });
-
-    it('should be true if monthOfBirth is not valid', function() {
-      // given
-      controller.set('dayOfBirth', '15');
-      controller.set('monthOfBirth', '99');
-      controller.set('yearOfBirth', '1999');
-
-      // when
-      const result = controller.get('isFormNotValid');
-
-      // then
-      expect(result).to.equal(true);
-    });
-
-    it('should be true if yearOfBirth is not valid', function() {
-      // given
-      controller.set('dayOfBirth', '15');
-      controller.set('monthOfBirth', '12');
-      controller.set('yearOfBirth', '99999');
-
-      // when
-      const result = controller.get('isFormNotValid');
-
-      // then
-      expect(result).to.equal(true);
-    });
-
-    it('should be false', function() {
-      // given
-      controller.set('firstName', 'Robert');
-      controller.set('lastName', 'Smith');
-      controller.set('dayOfBirth', '15');
-      controller.set('monthOfBirth', '12');
-      controller.set('yearOfBirth', '1999');
-
-      // when
-      const result = controller.get('isFormNotValid');
-
-      // then
-      expect(result).to.equal(false);
-    });
-  });
-
-  describe('#isDisabled', function() {
-
-    it('should be false if source is not external', function() {
-      // when
-      const result = controller.get('isDisabled');
-
-      // then
-      expect(result).to.equal(false);
-    });
-
-    it('should be true if source is external', function() {
-      // given
-      sessionStub.data.authenticated.source = 'external';
-
-      // when
-      const result = controller.get('isDisabled');
-
-      // then
-      expect(result).to.equal(true);
-    });
-  });
-
   describe('#attemptNext', function() {
-
-    beforeEach(function() {
-      controller.set('firstName', 'Robert');
-      controller.set('lastName', 'Smith');
-      controller.set('dayOfBirth', '10');
-      controller.set('monthOfBirth', '10');
-      controller.set('yearOfBirth', '2000');
-    });
-
-    it('should display an error on firstName', async function() {
-      // given
-      controller.set('firstName', ' ');
-
-      // when
-      await controller.actions.attemptNext.call(controller);
-
-      // then
-      sinon.assert.notCalled(storeStub.createRecord);
-      sinon.assert.notCalled(studentUserAssociation.save);
-      expect(controller.get('isLoading')).to.equal(false);
-      sinon.assert.notCalled(controller.transitionToRoute);
-      expect(controller.get('errorMessage')).to.equal(null);
-      expect(controller.get('validation.firstName')).to.equal('Votre prénom n’est pas renseigné.');
-    });
-
-    it('should display an error on lastName', async function() {
-      // given
-      controller.set('lastName', '');
-
-      // when
-      await controller.actions.attemptNext.call(controller);
-
-      // then
-      sinon.assert.notCalled(storeStub.createRecord);
-      sinon.assert.notCalled(studentUserAssociation.save);
-      expect(controller.get('isLoading')).to.equal(false);
-      sinon.assert.notCalled(controller.transitionToRoute);
-      expect(controller.get('errorMessage')).to.equal(null);
-      expect(controller.get('validation.lastName')).to.equal('Votre nom n’est pas renseigné.');
-    });
-
-    it('should display an error on dayOfBirth', async function() {
-      // given
-      controller.set('dayOfBirth', '99');
-
-      // when
-      await controller.actions.attemptNext.call(controller);
-
-      // then
-      sinon.assert.notCalled(storeStub.createRecord);
-      sinon.assert.notCalled(studentUserAssociation.save);
-      expect(controller.get('isLoading')).to.equal(false);
-      sinon.assert.notCalled(controller.transitionToRoute);
-      expect(controller.get('errorMessage')).to.equal(null);
-      expect(controller.get('validation.dayOfBirth')).to.equal('Votre jour de naissance n’est pas valide.');
-    });
-
-    it('should display an error on monthOfBirth', async function() {
-      // given
-      controller.set('monthOfBirth', '99');
-
-      // when
-      await controller.actions.attemptNext.call(controller);
-
-      // then
-      sinon.assert.notCalled(storeStub.createRecord);
-      sinon.assert.notCalled(studentUserAssociation.save);
-      expect(controller.get('isLoading')).to.equal(false);
-      sinon.assert.notCalled(controller.transitionToRoute);
-      expect(controller.get('errorMessage')).to.equal(null);
-      expect(controller.get('validation.monthOfBirth')).to.equal('Votre mois de naissance n’est pas valide.');
-    });
-
-    it('should display an error on yearOfBirth', async function() {
-      // given
-      controller.set('yearOfBirth', '99');
-
-      // when
-      await controller.actions.attemptNext.call(controller);
-
-      // then
-      sinon.assert.notCalled(storeStub.createRecord);
-      sinon.assert.notCalled(studentUserAssociation.save);
-      expect(controller.get('isLoading')).to.equal(false);
-      sinon.assert.notCalled(controller.transitionToRoute);
-      expect(controller.get('errorMessage')).to.equal(null);
-      expect(controller.get('validation.yearOfBirth')).to.equal('Votre année de naissance n’est pas valide.');
-    });
-
-    it('should display an error on all fields', async function() {
-      // given
-      controller.set('firstName', '');
-      controller.set('lastName', '');
-      controller.set('dayOfBirth', '');
-      controller.set('monthOfBirth', '');
-      controller.set('yearOfBirth', '');
-
-      // when
-      await controller.actions.attemptNext.call(controller);
-
-      // then
-      sinon.assert.notCalled(storeStub.createRecord);
-      sinon.assert.notCalled(studentUserAssociation.save);
-      expect(controller.get('isLoading')).to.equal(false);
-      sinon.assert.notCalled(controller.transitionToRoute);
-      expect(controller.get('errorMessage')).to.equal(null);
-      expect(controller.get('validation.firstName')).to.equal('Votre prénom n’est pas renseigné.');
-      expect(controller.get('validation.lastName')).to.equal('Votre nom n’est pas renseigné.');
-      expect(controller.get('validation.dayOfBirth')).to.equal('Votre jour de naissance n’est pas valide.');
-      expect(controller.get('validation.monthOfBirth')).to.equal('Votre mois de naissance n’est pas valide.');
-      expect(controller.get('validation.yearOfBirth')).to.equal('Votre année de naissance n’est pas valide.');
-    });
-
-    it('should be valid after an error', async function() {
-      // given
-      studentUserAssociation.save.resolves();
-      controller.set('yearOfBirth', '99');
-      await controller.actions.attemptNext.call(controller);
-
-      // when
-      controller.set('yearOfBirth', '2000');
-      await controller.actions.attemptNext.call(controller);
-
-      // then
-      sinon.assert.calledOnce(storeStub.createRecord);
-      sinon.assert.calledOnce(studentUserAssociation.save);
-      expect(controller.get('isLoading')).to.equal(false);
-      sinon.assert.calledWith(controller.transitionToRoute, 'campaigns.start-or-resume');
-      expect(controller.get('errorMessage')).to.equal(null);
-      expect(controller.get('validation.yearOfBirth')).to.equal(null);
-    });
 
     it('should associate user with student and redirect to campaigns.start-or-resume', async function() {
       // given
-      studentUserAssociation.save.resolves();
+      schoolingRegistration.save.resolves();
 
       // when
-      await controller.actions.attemptNext.call(controller);
+      await controller.actions.attemptNext.call(controller, schoolingRegistration);
 
       // then
-      sinon.assert.calledOnce(storeStub.createRecord);
-      sinon.assert.calledOnce(studentUserAssociation.save);
-      expect(controller.get('isLoading')).to.equal(false);
+      sinon.assert.calledOnce(schoolingRegistration.save);
       sinon.assert.calledWith(controller.transitionToRoute, 'campaigns.start-or-resume');
-      expect(controller.get('errorMessage')).to.equal(null);
     });
 
     it('should display a not found error', async function() {
       // given
-      studentUserAssociation.save.rejects({ errors: [{ status: '404' }] });
+      schoolingRegistration.save.rejects({ errors: [{ status: '404' }] });
 
       // when
-      await controller.actions.attemptNext.call(controller);
+      await controller.actions.attemptNext.call(controller, schoolingRegistration);
 
       // then
-      sinon.assert.calledOnce(studentUserAssociation.unloadRecord);
+      sinon.assert.calledOnce(schoolingRegistration.unloadRecord);
       expect(controller.get('errorMessage')).to.equal('Vous êtes un élève ? <br/> Vérifiez vos informations (prénom, nom et date de naissance) ou contactez un enseignant.<br/> <br/> Vous êtes un enseignant ? <br/> L‘accès à un parcours n‘est pas disponible pour le moment.');
       sinon.assert.notCalled(controller.transitionToRoute);
-      expect(controller.get('isLoading')).to.equal(false);
-    });
-
-    it('should associate user with student and redirect to campaigns.start-or-resume after failing', async function() {
-      // given
-      studentUserAssociation.save
-        .onFirstCall().rejects({ errors: [{ status: '404' }] })
-        .onSecondCall().resolves();
-
-      // when
-      // first
-      await controller.actions.attemptNext.call(controller);
-      // second
-      await controller.actions.attemptNext.call(controller);
-
-      // then
-      sinon.assert.calledTwice(storeStub.createRecord);
-      sinon.assert.calledTwice(studentUserAssociation.save);
-      expect(controller.get('isLoading')).to.equal(false);
-      sinon.assert.calledWith(controller.transitionToRoute, 'campaigns.start-or-resume');
-      expect(controller.get('errorMessage')).to.equal(null);
     });
 
     describe('When student is already reconciled in the same organization', async function() {
@@ -502,13 +65,13 @@ describe('Unit | Controller | campaigns/restricted/join', function() {
           };
           const expectedErrorMessage = 'Vous possédez déjà un compte Pix utilisé dans votre établissement scolaire, avec l‘adresse mail <br>j***@example.net.<br>Pour continuer, connectez-vous à ce compte ou demandez de l’aide à un enseignant.<br> (Code R31)';
 
-          studentUserAssociation.save.rejects({ errors: [error] });
+          schoolingRegistration.save.rejects({ errors: [error] });
 
           // when
           await controller.actions.attemptNext.call(controller);
 
           // then
-          sinon.assert.calledOnce(studentUserAssociation.unloadRecord);
+          sinon.assert.calledOnce(schoolingRegistration.unloadRecord);
           expect(controller.get('errorMessage')).to.equal(expectedErrorMessage);
           sinon.assert.notCalled(controller.transitionToRoute);
           expect(controller.get('isLoading')).to.equal(false);
@@ -528,13 +91,13 @@ describe('Unit | Controller | campaigns/restricted/join', function() {
             meta: { shortCode: 'R32', value: 'j***.h***2' }
           };
           const expectedErrorMessage = 'Vous possédez déjà un compte Pix utilisé dans votre établissement scolaire, avec l‘identifiant<br>j***.h***2.<br>Pour continuer, connectez-vous à ce compte ou demandez de l‘aide à un enseignant.<br> (Code R32)';
-          studentUserAssociation.save.rejects({ errors: [error] });
+          schoolingRegistration.save.rejects({ errors: [error] });
 
           // when
           await controller.actions.attemptNext.call(controller);
 
           // then
-          sinon.assert.calledOnce(studentUserAssociation.unloadRecord);
+          sinon.assert.calledOnce(schoolingRegistration.unloadRecord);
           expect(controller.get('errorMessage')).to.equal(expectedErrorMessage);
           sinon.assert.notCalled(controller.transitionToRoute);
           expect(controller.get('isLoading')).to.equal(false);
@@ -554,13 +117,13 @@ describe('Unit | Controller | campaigns/restricted/join', function() {
             meta: { shortCode: 'R33', value: undefined }
           };
           const expectedErrorMessage = 'Vous possédez déjà un compte Pix via l’ENT. Utilisez-le pour rejoindre le parcours.';
-          studentUserAssociation.save.rejects({ errors: [error] });
+          schoolingRegistration.save.rejects({ errors: [error] });
 
           // when
           await controller.actions.attemptNext.call(controller);
 
           // then
-          sinon.assert.calledOnce(studentUserAssociation.unloadRecord);
+          sinon.assert.calledOnce(schoolingRegistration.unloadRecord);
           expect(controller.get('errorMessage')).to.equal(expectedErrorMessage);
           sinon.assert.notCalled(controller.transitionToRoute);
           expect(controller.get('isLoading')).to.equal(false);
@@ -580,13 +143,13 @@ describe('Unit | Controller | campaigns/restricted/join', function() {
             meta: { shortCode: 'R33', value: undefined }
           };
           const expectedErrorMessage = 'Vous possédez déjà un compte Pix via l’ENT. Utilisez-le pour rejoindre le parcours.';
-          studentUserAssociation.save.rejects({ errors: [error] });
+          schoolingRegistration.save.rejects({ errors: [error] });
 
           // when
           await controller.actions.attemptNext.call(controller);
 
           // then
-          sinon.assert.calledOnce(studentUserAssociation.unloadRecord);
+          sinon.assert.calledOnce(schoolingRegistration.unloadRecord);
           expect(controller.get('errorMessage')).to.equal(expectedErrorMessage);
           sinon.assert.notCalled(controller.transitionToRoute);
           expect(controller.get('isLoading')).to.equal(false);
@@ -606,13 +169,13 @@ describe('Unit | Controller | campaigns/restricted/join', function() {
             meta: { shortCode: 'R33', value: undefined }
           };
           const expectedErrorMessage = 'Vous possédez déjà un compte Pix via l’ENT. Utilisez-le pour rejoindre le parcours.';
-          studentUserAssociation.save.rejects({ errors: [error] });
+          schoolingRegistration.save.rejects({ errors: [error] });
 
           // when
           await controller.actions.attemptNext.call(controller);
 
           // then
-          sinon.assert.calledOnce(studentUserAssociation.unloadRecord);
+          sinon.assert.calledOnce(schoolingRegistration.unloadRecord);
           expect(controller.get('errorMessage')).to.equal(expectedErrorMessage);
           sinon.assert.notCalled(controller.transitionToRoute);
           expect(controller.get('isLoading')).to.equal(false);
@@ -632,13 +195,13 @@ describe('Unit | Controller | campaigns/restricted/join', function() {
             meta: { shortCode: 'R33', value: undefined }
           };
           const expectedErrorMessage = 'Vous possédez déjà un compte Pix via l’ENT. Utilisez-le pour rejoindre le parcours.';
-          studentUserAssociation.save.rejects({ errors: [error] });
+          schoolingRegistration.save.rejects({ errors: [error] });
 
           // when
           await controller.actions.attemptNext.call(controller);
 
           // then
-          sinon.assert.calledOnce(studentUserAssociation.unloadRecord);
+          sinon.assert.calledOnce(schoolingRegistration.unloadRecord);
           expect(controller.get('errorMessage')).to.equal(expectedErrorMessage);
           sinon.assert.notCalled(controller.transitionToRoute);
           expect(controller.get('isLoading')).to.equal(false);
@@ -658,13 +221,13 @@ describe('Unit | Controller | campaigns/restricted/join', function() {
             meta: { shortCode: 'R32', value: 'j***.h***2' }
           };
           const expectedErrorMessage = 'Vous possédez déjà un compte Pix utilisé dans votre établissement scolaire, avec l‘identifiant<br>j***.h***2.<br>Pour continuer, connectez-vous à ce compte ou demandez de l‘aide à un enseignant.<br> (Code R32)';
-          studentUserAssociation.save.rejects({ errors: [error] });
+          schoolingRegistration.save.rejects({ errors: [error] });
 
           // when
           await controller.actions.attemptNext.call(controller);
 
           // then
-          sinon.assert.calledOnce(studentUserAssociation.unloadRecord);
+          sinon.assert.calledOnce(schoolingRegistration.unloadRecord);
           expect(controller.get('errorMessage')).to.equal(expectedErrorMessage);
           sinon.assert.notCalled(controller.transitionToRoute);
           expect(controller.get('isLoading')).to.equal(false);
@@ -688,13 +251,13 @@ describe('Unit | Controller | campaigns/restricted/join', function() {
             meta: { shortCode: 'R11', value: 'j***@example.net' }
           };
           const expectedErrorMessage = 'Vous possédez déjà un compte Pix avec l’adresse e-mail <br>j***@example.net<br>Pour continuer, connectez-vous à ce compte ou demandez de l’aide à un enseignant.<br>(Code R11)';
-          studentUserAssociation.save.rejects({ errors: [error] });
+          schoolingRegistration.save.rejects({ errors: [error] });
 
           // when
           await controller.actions.attemptNext.call(controller);
 
           // then
-          sinon.assert.calledOnce(studentUserAssociation.unloadRecord);
+          sinon.assert.calledOnce(schoolingRegistration.unloadRecord);
           expect(controller.get('errorMessage')).to.equal(expectedErrorMessage);
           sinon.assert.notCalled(controller.transitionToRoute);
           expect(controller.get('isLoading')).to.equal(false);
@@ -714,13 +277,13 @@ describe('Unit | Controller | campaigns/restricted/join', function() {
             meta: { shortCode: 'R12', value: 'j***.h***2' }
           };
           const expectedErrorMessage = 'Vous possédez déjà un compte Pix utilisé avec l’identifiant <br>j***.h***2<br>Pour continuer, connectez-vous à ce compte ou demandez de l’aide à un enseignant.<br>(Code R12)';
-          studentUserAssociation.save.rejects({ errors: [error] });
+          schoolingRegistration.save.rejects({ errors: [error] });
 
           // when
           await controller.actions.attemptNext.call(controller);
 
           // then
-          sinon.assert.calledOnce(studentUserAssociation.unloadRecord);
+          sinon.assert.calledOnce(schoolingRegistration.unloadRecord);
           expect(controller.get('errorMessage')).to.equal(expectedErrorMessage);
           sinon.assert.notCalled(controller.transitionToRoute);
           expect(controller.get('isLoading')).to.equal(false);
@@ -740,13 +303,13 @@ describe('Unit | Controller | campaigns/restricted/join', function() {
             meta: { shortCode: 'R13', value: undefined }
           };
           const expectedErrorMessage = 'Vous possédez déjà un compte Pix via l‘ENT dans un autre établissement scolaire.<br>Pour continuer, contactez un enseignant qui pourra vous donner l’accès à ce compte à l‘aide de Pix Orga.';
-          studentUserAssociation.save.rejects({ errors: [error] });
+          schoolingRegistration.save.rejects({ errors: [error] });
 
           // when
           await controller.actions.attemptNext.call(controller);
 
           // then
-          sinon.assert.calledOnce(studentUserAssociation.unloadRecord);
+          sinon.assert.calledOnce(schoolingRegistration.unloadRecord);
           expect(controller.get('errorMessage')).to.equal(expectedErrorMessage);
           sinon.assert.notCalled(controller.transitionToRoute);
           expect(controller.get('isLoading')).to.equal(false);
@@ -766,13 +329,13 @@ describe('Unit | Controller | campaigns/restricted/join', function() {
             meta: { shortCode: 'R13', value: undefined }
           };
           const expectedErrorMessage = 'Vous possédez déjà un compte Pix via l‘ENT dans un autre établissement scolaire.<br>Pour continuer, contactez un enseignant qui pourra vous donner l’accès à ce compte à l‘aide de Pix Orga.';
-          studentUserAssociation.save.rejects({ errors: [error] });
+          schoolingRegistration.save.rejects({ errors: [error] });
 
           // when
           await controller.actions.attemptNext.call(controller);
 
           // then
-          sinon.assert.calledOnce(studentUserAssociation.unloadRecord);
+          sinon.assert.calledOnce(schoolingRegistration.unloadRecord);
           expect(controller.get('errorMessage')).to.equal(expectedErrorMessage);
           sinon.assert.notCalled(controller.transitionToRoute);
           expect(controller.get('isLoading')).to.equal(false);
@@ -792,13 +355,13 @@ describe('Unit | Controller | campaigns/restricted/join', function() {
             meta: { shortCode: 'R13', value: undefined }
           };
           const expectedErrorMessage = 'Vous possédez déjà un compte Pix via l‘ENT dans un autre établissement scolaire.<br>Pour continuer, contactez un enseignant qui pourra vous donner l’accès à ce compte à l‘aide de Pix Orga.';
-          studentUserAssociation.save.rejects({ errors: [error] });
+          schoolingRegistration.save.rejects({ errors: [error] });
 
           // when
           await controller.actions.attemptNext.call(controller);
 
           // then
-          sinon.assert.calledOnce(studentUserAssociation.unloadRecord);
+          sinon.assert.calledOnce(schoolingRegistration.unloadRecord);
           expect(controller.get('errorMessage')).to.equal(expectedErrorMessage);
           sinon.assert.notCalled(controller.transitionToRoute);
           expect(controller.get('isLoading')).to.equal(false);
@@ -818,13 +381,13 @@ describe('Unit | Controller | campaigns/restricted/join', function() {
             meta: { shortCode: 'R13', value: undefined }
           };
           const expectedErrorMessage = 'Vous possédez déjà un compte Pix via l‘ENT dans un autre établissement scolaire.<br>Pour continuer, contactez un enseignant qui pourra vous donner l’accès à ce compte à l‘aide de Pix Orga.';
-          studentUserAssociation.save.rejects({ errors: [error] });
+          schoolingRegistration.save.rejects({ errors: [error] });
 
           // when
           await controller.actions.attemptNext.call(controller);
 
           // then
-          sinon.assert.calledOnce(studentUserAssociation.unloadRecord);
+          sinon.assert.calledOnce(schoolingRegistration.unloadRecord);
           expect(controller.get('errorMessage')).to.equal(expectedErrorMessage);
           sinon.assert.notCalled(controller.transitionToRoute);
           expect(controller.get('isLoading')).to.equal(false);
@@ -844,13 +407,13 @@ describe('Unit | Controller | campaigns/restricted/join', function() {
             meta: { shortCode: 'R12', value: 'j***.h***2' }
           };
           const expectedErrorMessage = 'Vous possédez déjà un compte Pix utilisé avec l’identifiant <br>j***.h***2<br>Pour continuer, connectez-vous à ce compte ou demandez de l’aide à un enseignant.<br>(Code R12)';
-          studentUserAssociation.save.rejects({ errors: [error] });
+          schoolingRegistration.save.rejects({ errors: [error] });
 
           // when
           await controller.actions.attemptNext.call(controller);
 
           // then
-          sinon.assert.calledOnce(studentUserAssociation.unloadRecord);
+          sinon.assert.calledOnce(schoolingRegistration.unloadRecord);
           expect(controller.get('errorMessage')).to.equal(expectedErrorMessage);
           sinon.assert.notCalled(controller.transitionToRoute);
           expect(controller.get('isLoading')).to.equal(false);
